genXXX.py: default to not clean up rootfs
Previously LAT defaults to clean up rootfs unless '--no-clean' is
supplied. This is not very good because if something goes wrong,
user has to re-run the command with '--no-clean' option to enable
further debugging, not to mention errors that are not 100% reproducible.

We need to change the behavior now, keeping the rootfs by default.
This has the advange of keeping the scene when something goes wrong.

We still support the '--no-clean' option, but with a different meaning.
It now means "not clean up previously generated rootfs". In this way,
people could still do incremental rootfs generation.

The previous behavior of cleaning up rootfs is mainly about solving the
pseudo mismatch problem. Unfortunately it's also not quite correct, because
the atexit functions are not always called, for example in case of SIGKILL
or SIGTERM. This means current behaviour does not ensure rootfs cleanup
before creating a new rootfs.

In summary, changing to clean rootfs at prepration stage has two advantages:
1. Ensure the rootfs and pseudo dirs are cleaned before doing rootfs by default.
2. Keep the scene if anything goes wrong.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>