genimage: use minbase variant for debootstrap
By default debootstrap creates a base Debian installation with all
packages of a priority required and important. This includes packages
such as systemd, systemd-sys, and systemd-timesyncd. This creates a
race condition when debootstrap is complete and then the starlingx
packages are updated after. This is most notable while trying to
install ntpd and systemd-timesyncd is already installed.

When using the minbase variant debootstrap will only install the
required packages in order to use a Debian system. This includes
apt, base-files, base-passwd, bash, and etc.

https://review.opendev.org/c/starlingx/tools/+/827164

Suggested-by: Church, Robert <Robert.Church@windriver.com>

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>