genimage: improve yaml input param checking
While checking the yaml input parameters that whether
the file exist or not, if the file locates in target rootfs,
skip checking and return true. Because the param checking
is earlier than do_rootfs, it does not make sense to
check file existence in target rootfs at param checking time.

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>