fullmetalupdate.bbclass: HAWKBIT_PASSWD_FILE -> HAWKBIT_USER_PASSWORD
Issue: LINCD-8811

Read the password from a file isn't good for OOBE, and doesn't improve the
security, so set and read if from the var HAWKBIT_USER_PASSWORD to make it
simple, and warn it if the default password is being used.

(LOCAL REV: NOT UPSTREAM) -- Will send it later with other fixes.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>