lat installer: fix conflict of multiple installer ISO image
If there are multiple disks, and each disk has been burned one installer
ISO image, originally the label of installer ISO image is hardcoded as
`instboot', in this situation, the installer could not detect the
right disk of ISO image.

Unique the label of installer ISO image, append an 8-character
short unique random string to "instboot-iso", pass it as boot params
to installer script. Then the installer could detect and mount the
right ISO image for installation.

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>