genimage: customize the mini initramfs for subcloud
Issue: CGTS-37459

For Redfish Subcloud Install it needs a small customizable initramfs
less than 80M.
Now we add a task to create this mini initramfs and compress it with
lzma format.
The detailed steps are as follows:
 1.Decompress the normal initramfs generated by lat with the yaml file.
 2.Drop rt kernel modules, __pycache__ directories, local entries...
 3.Generate the new initrd and compress it using xz command

Install the mini initrd into the /var/miniboot direcotry for rootfs,
and WRCP team will use it after the controller-0 node installation
during Redfish Subcloud boot.

Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>