genimage: remove crda from bcm-2xxx-rpi4.yaml
CRDA is no longer needed as of kernel v4.15 since commit 007f6c5e6eb45
("cfg80211: support loading regulatory database as firmware file") added
support to use the kernel's firmware request API which looks for the
firmware on /lib/firmware. Because of this CRDA is legacy software for
older kernels, not install it any more.

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>