lat installer: print a warn to install on non recommend disk for non intel-x86-64 bsp
For ARM BSPs, the recommend disk is the installer image device,
if user install on non recommend disk, the install may fail or
do re-install repeatedly. Print a warn to let user know

...
    NAME   VENDOR      SIZE MODEL            TYPE LABEL
0 - sda  QEMU        640M QEMU HARDDISK    disk
1 - vda  0x554d4551   16G                  disk
2 - sdb  QEMU         32G QEMU HARDDISK    disk
B - Reboot
R - Refresh

Press any other key to install on: 1?
Or select disk to format and install: 2

WARNING: You choose to install on *sdb* which is not the suggested disk *vda*
WARNING: It may fail to install and re-install repeatedly, or failed to boot after installed.
ERASE /dev/sdb (y/n)
...

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>