lat installer: improve the interactive disk selection
Press Enter to install on sdX or Esc to selelct a different device
While Esc, ask to input invalid index of disk

1. Press Enter
...
    NAME   VENDOR    SIZE MODEL            TYPE LABEL
0 - sda  QEMU       32G QEMU HARDDISK    disk
1 - sdc  QEMU       64G QEMU HARDDISK    disk
B - Reboot
R - Refresh

Press Enter to to install on sdc or Esc to selelct a different device?
Choose to install on disk: 1 (sdc  QEMU       64G QEMU HARDDISK    disk)

ERASE /dev/sdc (y/n) y
...

2. Press Esc
...
    NAME   VENDOR    SIZE MODEL            TYPE LABEL
0 - sda  QEMU       32G QEMU HARDDISK    disk
1 - sdc  QEMU       64G QEMU HARDDISK    disk
B - Reboot
R - Refresh

Press Enter to to install on sdc or Esc to selelct a different device?
Select disk to format and install: a
The number 0 ~ 1 is required: 3
The number 0 ~ 1 is required: 0

The disk /dev/sda is not empty, ERASE /dev/sda (y/n)

Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>